# swagger-output_crudGrrop.json
{
  "swagger": "2.0",
  "info": {
    "title": "Node.js API Documentation",
    "description": "Auto-generated API routes",
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/projects/{projectId}/bv-items": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sourceJobTypeId": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "keterangan": {
                  "example": "any"
                },
                "paymentUnit": {
                  "example": "any"
                },
                "groupId": {
                  "example": "any"
                },
                "ecommerceLink": {
                  "example": "any"
                },
                "breakdowns": {
                  "example": "any"
                },
                "parentBvItemId": {
                  "example": "any"
                },
                "isHeaderOnly": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bv-items/{id}": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "sourceJobTypeId": {
                  "example": "any"
                },
                "name": {
                  "example": "any"
                },
                "keterangan": {
                  "example": "any"
                },
                "paymentUnit": {
                  "example": "any"
                },
                "groupId": {
                  "example": "any"
                },
                "ecommerceLink": {
                  "example": "any"
                },
                "breakdowns": {
                  "example": "any"
                },
                "parentBvItemId": {
                  "example": "any"
                },
                "isHeaderOnly": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bv-items/{id}/link-to-rab": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "rabUnitPrice": {
                  "example": "any"
                },
                "groupId": {
                  "example": "any"
                },
                "category": {
                  "example": "any"
                },
                "reference": {
                  "example": "any"
                },
                "overhead": {
                  "example": "any"
                },
                "components": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/bv-items/{id}/sync": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  }
}